requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# API Clients
tavily-python>=0.3.0
mistralai>=0.1.0

# Environment & Configuration
python-dotenv>=1.0.0

# HTTP & Utilities
httpx>=0.25.0
pydantic>=2.5.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Production Server (optional - for deployment)
gunicorn>=21.0.0
